css: accumulate copies of fixed and absolute divs

+++ src/drawlist.c
@@ -551,7 +551,7 @@ const CtxEntry *ctx_get_drawlist (Ctx *ctx, int *count)
 
 void ctx_drawlist_force_count (Ctx *ctx, int count)
 {
-   if (ctx->drawlist.count < count)
+   if ((int)ctx->drawlist.count < count)
      return;
    ctx->drawlist.count = count;
 }
